src: refactored lib/utility
* removed unused utilityMath
* removed some unused functions from utility
* moved time related utility functions to TimeStamp
* moved path objects from utility to app/paths
* moved objects in lib to lib/app
* removed duplicate confirm methods in DialogView
* moved utilityFile to utility/file
* removed uint typedef and replaced with size_t everywhere
* removed boost header in utilityUuid

+++ src/lib/component/view/helper/CodeSnippetParams.h
@@ -16,8 +16,8 @@ struct CodeSnippetParams
 	static bool sort(const CodeSnippetParams& a, const CodeSnippetParams& b);
 	static bool sortById(const CodeSnippetParams& a, const CodeSnippetParams& b);
 
-	uint startLineNumber;
-	uint endLineNumber;
+	size_t startLineNumber;
+	size_t endLineNumber;
 
 	std::wstring title;
 	std::wstring footer;
