+++ src/lib_gui/qt/element/code/QtCodeFileSingle.h
@@ -38,7 +38,7 @@ public:
 
 	void requestFileContent(const FilePath& filePath) override;
 	bool requestScroll(
-		const FilePath& filePath, uint lineNumber, Id locationId, bool animated, ScrollTarget target) override;
+		const FilePath& filePath, size_t lineNumber, Id locationId, bool animated, ScrollTarget target) override;
 
 	void updateFiles() override;
 	void showContents() override;
